requirements.txt
biopython==1.76
certifi==2019.11.28
click==7.1.1
cycler==0.10.0
dash==1.9.1
dash-core-components==1.8.1
dash-html-components==1.0.2
dash-renderer==1.2.4
dash-table==4.6.1
Flask==1.1.1
Flask-Compress==1.4.0
future==0.18.2
gunicorn==20.0.4
ipykernel==5.1.4
itsdangerous==1.1.0
Jinja2==2.11.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.2.1
numpy==1.18.2
pandas==1.0.3
plotly==4.5.4
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
retrying==1.3.3
scipy==1.4.1
seaborn==0.10.0
six==1.14.0
Werkzeug==1.0.0
